// src/generated-code/angular/horseless-odata-content-workspace/msgraph/microsoft/graph/keycredential.complex.ts
//#region ODataApiGen ODataImports
import {
  Duration
} from 'angular-odata';//#endregion

//#region ODataApiGen Imports
//#endregion

export interface KeyCredential {
  //#region ODataApiGen Properties
  customKeyIdentifier?: ArrayBuffer;
  displayName?: string;
  endDateTime?: Date;
  keyId?: string;
  startDateTime?: Date;
  type?: string;
  usage?: string;
  key?: ArrayBuffer;
  //#endregion
}